Main/VRDEServerImpl.cpp: Signed/unsigned conversion issues. bugref:9790
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85255 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/VRDEServerImpl.cpp
@@ -285,13 +285,13 @@ static int i_portParseNumber(uint16_t *pu16Port, const char *pszStart, const cha
      *       only digits and pszEnd points to the char after last
      *       digit.
      */
-    size_t cch = pszEnd - pszStart;
+    size_t cch = (size_t)(pszEnd - pszStart);
     if (cch > 0 && cch <= 5) /* Port is up to 5 decimal digits. */
     {
         unsigned uPort = 0;
         while (pszStart != pszEnd)
         {
-            uPort = uPort * 10 + *pszStart - '0';
+            uPort = uPort * 10 + (unsigned)(*pszStart - '0');
             pszStart++;
         }
